\name{VennPlots-package}
\alias{VennPlots-package}
\alias{VennPlots}
\docType{package}
\title{
\packageTitle{VennPlots}
}
\description{
\packageDescription{VennPlots}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VennPlots}
\packageIndices{VennPlots}

}
\author{
\packageAuthor{VennPlots}

Maintainer: \packageMaintainer{VennPlots}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
