\name{VennPlots-package}
\alias{VennPlots-package}
\alias{VennPlots}
\docType{package}
\title{
\packageTitle{VennPlots}
}
\description{
\packageDescription{VennPlots}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VennPlots}
\packageIndices{VennPlots}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{VennPlots}

Maintainer: \packageMaintainer{VennPlots}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
